@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 import { ModelNewService } from './model-new.service';
 import {Router} from "@angular/router"
 import { Model } from '../types';
@@ -12,9 +12,9 @@ import {MatSnackBar} from '@angular/material/snack-bar';
   providers: [ModelNewService]
 })
 export class ModelNewComponent {
-  newModelForm = new FormGroup({
-    nameControl: new FormControl("", [Validators.required, noInValidCharacatersValidator]),
-    descriptionControl: new FormControl(""),
+  newModelForm = new UntypedFormGroup({
+    nameControl: new UntypedFormControl("", [Validators.required, noInValidCharacatersValidator]),
+    descriptionControl: new UntypedFormControl(""),
   })
   
 
@@ -38,7 +38,7 @@ export class ModelNewComponent {
 }
 
 
-function noInValidCharacatersValidator(control: FormControl) {
+function noInValidCharacatersValidator(control: UntypedFormControl) {
   const invalidCharacter = '<>" {}|\\^`';
 
   for (const c of control?.value.split("")) {
