@@ -1,5 +1,3 @@
-from rdflib import Graph
-
 from buildingmotif.dataclasses import Library
 
 
@@ -27,78 +25,52 @@ def test_get_all_libraries(client, building_motif):
 
 
 def test_get_all_shapes(client, building_motif):
-    # Setup
     my_library = Library.create("my_library")
-    shape_graph_ttl = """
-    @prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
-    @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
-    @prefix skos: <http://www.w3.org/2004/02/skos/core#> .
-    @prefix brick: <https://brickschema.org/schema/Brick#> .
+    shape_collection = my_library.get_shape_collection()
+    shape_collection.graph.parse(
+        data="""
+    @prefix owl: <http://www.w3.org/2002/07/owl#> .
     @prefix sh: <http://www.w3.org/ns/shacl#> .
-    @prefix : <urn:ex/> .
-    :my_shape a sh:NodeShape ;
-        sh:targetClass brick:Entity ;
-        rdfs:label "my shape label" ;
-        skos:description "my shape description" .
-    :my_other_shape a sh:NodeShape ;
-        sh:targetClass brick:Entity ;
-        rdfs:label "my shape other label" ;
-        skos:description "my shape other description" .
-    """
-    g = Graph()
-    g.parse(data=shape_graph_ttl)
-    my_library.get_shape_collection().add_graph(g)
+    @prefix : <urn:model#> .
+    @prefix bmotif: <https://nrel.gov/BuildingMOTIF#> .
+    @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
-    shape_graph_ttl = """
-    @prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
-    @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
-    @prefix skos: <http://www.w3.org/2004/02/skos/core#> .
-    @prefix brick: <https://brickschema.org/schema/Brick#> .
-    @prefix sh: <http://www.w3.org/ns/shacl#> .
-    @prefix : <urn:ex/> .
-    :your_shape a sh:NodeShape ;
-        sh:targetClass brick:Entity ;
-        rdfs:label "your shape label" ;
-        skos:description "your shape description" .
+    :shape1 a owl:Class, bmotif:Definition_Type ;
+        rdfs:label "a";
+    .
+
+    :shape2 a owl:Class, bmotif:Sequence_Of_Operations ;
+        rdfs:label "b";
+    .
+
+    :shape3 a owl:Class, bmotif:Analytics_Application ;
+        rdfs:label "c";
+    .
     """
-    g = Graph()
-    g.parse(data=shape_graph_ttl)
-    your_library = Library.create("your_library")
-    your_library.get_shape_collection().add_graph(g)
+    )
 
     # Act
     results = client.get("/libraries/shapes")
 
-    # Assert
-    assert results.status_code == 200
-    expected = sorted(
-        [
+    assert results.json == {
+        "https://nrel.gov/BuildingMOTIF#Analytics_Application": [
             {
+                "shape_uri": "urn:model#shape3",
                 "library_name": "my_library",
-                "library_id": my_library.id,
-                "uri": "urn:ex/my_shape",
-                "label": "my shape label",
-                "description": "my shape description",
-            },
-            {
-                "library_name": "my_library",
-                "library_id": my_library.id,
-                "uri": "urn:ex/my_other_shape",
-                "label": "my shape other label",
-                "description": "my shape other description",
-            },
-            {
-                "library_name": "your_library",
-                "library_id": your_library.id,
-                "uri": "urn:ex/your_shape",
-                "label": "your shape label",
-                "description": "your shape description",
+                "shape_collection_id": shape_collection.id,
+                "label": "c",
             },
         ],
-        key=lambda x: x["uri"],
-    )
-    actual = sorted(results.json, key=lambda x: x["uri"])
-    assert expected == actual
+        "https://nrel.gov/BuildingMOTIF#Sequence_Of_Operations": [
+            {
+                "shape_uri": "urn:model#shape2",
+                "library_name": "my_library",
+                "shape_collection_id": shape_collection.id,
+                "label": "b",
+            },
+        ],
+        "https://nrel.gov/BuildingMOTIF#System_Specification": [],
+    }
 
 
 def test_get_library(client, building_motif):
